+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.4
+* JHijriPicker Fourth Release
+* Merge `showJDateTimePicker` with `showJHijriPicker`
 ## 0.0.3
 * JHijriPicker Third Release
 * Remove `requird WidgetType`